+++ tests/Testing/Testing/drawings/WallsDrawingObject.hpp
@@ -12,5 +12,5 @@ class WallsDrawingObject : public DrawingObject {
     public:
         WallsDrawingObject(int dimension, TextureAsset* textureAsset, float z,
                                 const char* vertexShaderPath, const char* fragmentShaderPath);
-        void initVerticesAndIndices(int i, int j, int dimension, float** map, const rapidjson::Document& wallsInfo);
+        void initVerticesAndIndices(int i, int j, int dimension, int** map, const rapidjson::Document& wallsInfo);
 };
