+++ tests/Testing/Testing/utils/TextureUtils.hpp
@@ -13,5 +13,5 @@ class TextureUtils {
 
 public:
 	static TextureAsset loadTextureAsset(const char* path, int numRows, int numCols, int vertexBufferSize, int verticesPerQuad, int indicesPerQuad);
-    static void initMap(int dimension, float** map, float** wallMap, PlatformsDrawingObject* platforms, WallsDrawingObject* walls, const rapidjson::Document& wallsInfo);
+    static void initMap(int dimension, int** map, int** wallMap, PlatformsDrawingObject* platforms, WallsDrawingObject* walls, const rapidjson::Document& wallsInfo);
 };
